+++ app/src/main/java/com/arctouch/codechallenge/features/home/HomeView.kt
@@ -8,5 +8,5 @@ interface HomeView: BaseView {
 
     fun showMovies(movies: MutableList<Movie>, pages: Int)
     fun showNextPage(movies: MutableList<Movie>)
-    fun callMovieDetailsActivity(movieId: Int)
+    fun callMovieDetailsActivity(movieId: Long)
 }
